* Added proper preference provider (based on LiteDb) with unit tests ** involved some refactoring in the database unit test * Bug fixes in volume management * implemented default VirtualVolumeService

+++ Server/soundbox/playback/system/IVirtualVolumeService.cs
@@ -24,7 +24,7 @@ namespace Soundbox
         /// Returns the current virtual system volume.
         /// </summary>
         /// <returns></returns>
-        Task<int> GetVolume();
+        Task<double> GetVolume();
 
         /// <summary>
         /// Register a new playback service instance with the volume service.
